+++ FlaskApp/commands/event_commands.py
@@ -8,10 +8,8 @@ from flask import flash
 import pygame
 from pymavlink import mavutil
 from threading import Lock, Thread
-from .upload_mission import auto, upload_mission, set_return, start_mission
+
 from FlaskApp.extensions import socketio, rover_connection
-
-
 
 
 
@@ -107,7 +105,6 @@ def stop_mission():
         print("mission is paused and stopped")
 
 
-
 def reset_mission():
 
     if rover_connection:
@@ -115,99 +112,8 @@ def reset_mission():
         print("Clearing all mission items")
         rover_connection.mav.mission_clear_all_send(rover_connection.target_system, rover_connection.target_component)
 
-        socketio.emit("messages", {"message": "Mission is Reset"})
-        print("mission is reset.")
-
-
-
-def mission():
-    class mission_item:
-        def __init__(self, seq, current, x, y, z):
-            self.seq = seq
-            self.frame = mavutil.mavlink.MAV_FRAME_GLOBAL_RELATIVE_ALT
-            self.command = mavutil.mavlink.MAV_CMD_NAV_WAYPOINT
-            self.current = current
-            self.auto = 1
-            self.param1 = 0.0
-            self.param2 = 2.00
-            self.param3 = 20.00
-            self.param4 = math.nan #idk
-            self.param5 = x
-            self.param6 = y 
-            self.param7 = z
-            self.mission_type = 0
-
-
-    print("-- Program Started")
-    socketio.emit("messages", {"message": "Program Started"})
-
-
-    while(rover_connection.target_system == 0):
-        print("-- Checking Heartbeat")
-        socketio.emit("messages", {"message": "Checking Heart Beat"})
-
-        rover_connection.wait_heartbeat()
-        print(" -- heatbeat from system (system %u component %u)" % (rover_connection.target_system, rover_connection.target_component))
-        socketio.emit("messages", {"message": f"heatbeat from system {rover_connection.target_system, rover_connection.target_component}"})
-
-
-    socketio.emit("messages", {"message": "Creating Way Points"})
-    mission_waypoints = []
-
-    mission_waypoints.append(mission_item(0, 0, 59.000000, 1, 0))
-    mission_waypoints.append(mission_item(1, 0, 31.55599420, -84.16967420, 0))
-    mission_waypoints.append(mission_item(2, 0, 31.55608680, -84.16967350, 0))
-
-    upload_mission(rover_connection, mission_waypoints)
-
-    print(" -- All Waypoints Created and Uploaded")
-    socketio.emit("messages", {"message": "All Waypoints Created and Uploaded"})
-
-
-
-def home_mission():
-
-    class mission_item:
-        def __init__(self, seq, current, x, y, z):
-            self.seq = seq
-            self.frame = mavutil.mavlink.MAV_FRAME_GLOBAL_RELATIVE_ALT
-            self.command = mavutil.mavlink.MAV_CMD_NAV_WAYPOINT
-            self.current = current
-            self.auto = 1
-            self.param1 = 0.0
-            self.param2 = 2.00
-            self.param3 = 20.00
-            self.param4 = math.nan #idk
-            self.param5 = x
-            self.param6 = y 
-            self.param7 = z
-            self.mission_type = 0
-
-
-    print("-- Program Started")
-    socketio.emit("messages", {"message": "Program Started"})
-
-
-    while(rover_connection.target_system == 0):
-        print("-- Checking Heartbeat")
-        socketio.emit("messages", {"message": "Checking Heart Beat"})
-
-        rover_connection.wait_heartbeat()
-        print(" -- heatbeat from system (system %u component %u)" % (rover_connection.target_system, rover_connection.target_component))
-        socketio.emit("messages", {"message": f"heatbeat from system {rover_connection.target_system, rover_connection.target_component}"})
-
-
-    socketio.emit("messages", {"message": "Creating Way Points"})
-    mission_waypoints = []
-
-    mission_waypoints.append(mission_item(0, 0, 59.000000, 1, 0))
-    mission_waypoints.append(mission_item(1, 0, 31.55599420, -84.16967420, 0))
-    mission_waypoints.append(mission_item(2, 0, 31.55608680, -84.16967350, 0))
-
-    upload_mission(rover_connection, mission_waypoints)
-
-    print(" -- All Waypoints Created and Uploaded HOME MISSION")
-    socketio.emit("messages", {"message": "All Waypoints Created and Uploaded"})
+        socketio.emit("messages", {"message": "Mission is RESET"})
+        print("mission is RESET.")
