+++ FlaskApp/paths/bus_ramp_waypoints
@@ -1,0 +1,8 @@
+QGC WPL 110
+0	1	0	16	0	0	0	0	31.5562697	-84.1704842	59.540000	1
+1	0	3	16	0.00000000	0.00000000	0.00000000	0.00000000	31.55627590	-84.17048360	100.000000	1
+2	0	3	16	0.00000000	0.00000000	0.00000000	0.00000000	31.55627140	-84.17072500	100.000000	1
+3	0	3	16	0.00000000	0.00000000	0.00000000	0.00000000	31.55626450	-84.17096640	100.000000	1
+4	0	3	16	0.00000000	0.00000000	0.00000000	0.00000000	31.55626450	-84.17096700	100.000000	1
+5	0	3	16	0.00000000	0.00000000	0.00000000	0.00000000	31.55627080	-84.17072560	100.000000	1
+6	0	3	16	0.00000000	0.00000000	0.00000000	0.00000000	31.55627650	-84.17048360	100.000000	1
