+++ FlaskApp/routes.py
@@ -6,4 +6,12 @@ main = Blueprint("main", __name__)
 
 @main.route("/")
 def index():
-    return render_template("example.html")
+    return render_template("home.html")
+
+@main.route("/home.html")
+def home():
+    return render_template("home.html")
+
+@main.route("/help.html")
+def help():
+    return render_template("help.html")
